{{cookiecutter.repo_name}}/requirements.txt
Click>=6.0
Flask-Migrate
Flask-SQLAlchemy
flask
pystatuschecker
sqlalchemy
ujson
ultra-config
webtest
pylint
colorlog
